requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
requests
python-dotenv
